Add some declarations for 10.5 compatibility, mostly for avoiding compiler warnings when linking against the 10.5 SDK.

+++ SKTransitionController.m
@@ -492,7 +492,7 @@ BOOL CoreGraphicsServicesTransitionsDefined() {
 
 - (void)prepareOpenGL {
     // Enable beam-synced updates.
-    long parm = 1;
+    GLint parm = 1;
     [[self openGLContext] setValues:&parm forParameter:NSOpenGLCPSwapInterval];
     
     // Make sure that everything we don't need is disabled.
